added encryption

+++ include/shifre.h
@@ -1,8 +1,7 @@
 #ifndef C_CHAT_SHIFRE_H
 #define C_CHAT_SHIFRE_H
 
-void encrypt(char *msg, int *buffer, int e, int n);
-void decrypt(int *encMsg, size_t encMsgSize,char *buffer, int d, int n);
-
+void encrypt(char *msg, long *buffer, int e, int n);
+void decrypt(long *encMsg, size_t encMsgSize,char *buffer, int d, int n);
 
 #endif
